requirements.txt
huggingface_hub>=0.34.4
pyyaml>=6.0.2
pillow>=11.1.0
numpy>=2.1.1
diffusers>=0.35.0
transformers>=4.55.0
accelerate>=1.10.0
safetensors>=0.5.3
lpips>=0.1.4
scikit-image>=0.25.2
pytest>=8.4.1
ruff>=0.12.0
